requirements.txt
loguru
openai
python-dotenv
loguru
streamlit
gradio